refactor(Gamer) : CardDeck 조합
